# src/SpaTemplate.Core/FacultyContext/Models/CourseForManipulationDto.cs
// -----------------------------------------------------------------------
// <copyright file="CourseForManipulationDto.cs" company="Piotr Xeinaemm Czech">
// Copyright (c) Piotr Xeinaemm Czech. All rights reserved.
// Licensed under the MIT license. See LICENSE file in the project root for full license information.
// </copyright>
// -----------------------------------------------------------------------

namespace SpaTemplate.Core.FacultyContext
{
    public class CourseForManipulationDto
    {
        public virtual string Description { get; set; }

        public string Title { get; set; }
    }
}