# src/SpaTemplate.Infrastructure.Api/wwwroot/v1/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v{version}/people/{studentId}/courses": {
      "post": {
        "tags": [
          "Courses"
        ],
        "summary": " ",
        "operationId": "Courses_CreateCourseForStudent",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 3
          }
        ],
        "requestBody": {
          "x-name": "courseForCreationDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseForCreationDto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "406": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": ""
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": " "
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelStateDictionary"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Courses"
        ],
        "summary": " ",
        "operationId": "Courses_GetCoursesForStudent",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "Accept",
            "in": "header",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 3
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 4
          }
        ],
        "requestBody": {
          "x-name": "parameters",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseParameters"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "406": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": ""
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": " "
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ModelStateDictionary": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "root": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ModelStateEntry"
              }
            ]
          },
          "maxAllowedErrors": {
            "type": "integer",
            "format": "int32"
          },
          "hasReachedMaxErrors": {
            "type": "boolean"
          },
          "errorCount": {
            "type": "integer",
            "format": "int32"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "keys": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "values": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelStateEntry"
            }
          },
          "isValid": {
            "type": "boolean"
          },
          "validationState": {
            "$ref": "#/components/schemas/ModelValidationState"
          },
          "item": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ModelStateEntry"
              }
            ]
          }
        }
      },
      "ModelStateEntry": {
        "type": "object",
        "x-abstract": true,
        "additionalProperties": false,
        "properties": {
          "rawValue": {
            "nullable": true
          },
          "attemptedValue": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ModelErrorCollection"
              }
            ]
          },
          "validationState": {
            "$ref": "#/components/schemas/ModelValidationState"
          }
        }
      },
      "ModelErrorCollection": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/ModelError"
        }
      },
      "ModelError": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "exception": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Exception"
              }
            ]
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Exception": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          },
          "innerException": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Exception"
              }
            ]
          },
          "stackTrace": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ModelValidationState": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Unvalidated",
          "Invalid",
          "Valid",
          "Skipped"
        ],
        "enum": [
          0,
          1,
          2,
          3
        ]
      },
      "CourseForCreationDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CourseForManipulationDto"
          },
          {
            "type": "object",
            "additionalProperties": false
          }
        ]
      },
      "CourseForManipulationDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CourseParameters": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "fields": {
            "type": "string",
            "nullable": true
          },
          "orderBy": {
            "type": "string",
            "nullable": true
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "searchQuery": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}